% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_DataAccess.R
\name{DataAccess}
\alias{DataAccess}
\title{Data Access}
\usage{
DataAccess(echo = NULL, csvFilePath = NULL)
}
\arguments{
\item{echo}{echo}

\item{csvFilePath}{file path for the excel file. csv file needs to be encoded as ANSI on Window OS.}
}
\description{
Data Access
}
\details{
Data Access
}
