% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_LCTM.R
\name{LCTMLinear}
\alias{LCTMLinear}
\title{LCTM}
\usage{
LCTMLinear(seed_num = 100, sample = NULL)
}
\arguments{
\item{seed_num}{seed_num}

\item{sample}{echo sample_data}

\item{fileName}{fileName}
}
\description{
LCTM
}
\details{
LCTM
}
