% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_LCTM.R
\name{LCTMLinearAddCV}
\alias{LCTMLinearAddCV}
\title{LCTMLinearAddCV}
\usage{
LCTMLinearAddCV(seed_num = 100, sample = NULL)
}
\arguments{
\item{seed_num}{seed_num}

\item{sample}{echo sample_data}
}
\description{
LCTMLinearAddCV
}
\details{
LCTMLinearAddCV
}
