% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_LCTM.R
\name{LCTMPlot}
\alias{LCTMPlot}
\title{LCTMPlot}
\usage{
LCTMPlot(sample, fileName = "test")
}
\arguments{
\item{sample}{echo sample_data}

\item{fileName}{fileName}
}
\description{
LCTMPlot
}
\details{
LCTMPlot
}
